+++ sdk/metric/internal/aggregate/filtered_reservoir.go
@@ -22,15 +22,20 @@ type FilteredExemplarReservoir[N int64 | float64] interface {
 	// The passed ctx needs to contain any baggage or span that were active
 	// when the measurement was made. This information may be used by the
 	// Reservoir in making a sampling decision.
-	Offer(ctx context.Context, val N, attr []attribute.KeyValue)
+	Offer(ctx context.Context, val N, lazySet attribute.LazyFilteredSet)
 	// Collect returns all the held exemplars in the reservoir.
 	Collect(dest *[]exemplar.Exemplar)
+}
+
+type lazyReservoir interface {
+	OfferLazy(context.Context, time.Time, exemplar.Value, attribute.LazyFilteredSet)
 }
 
 // filteredExemplarReservoir handles the pre-sampled exemplar of measurements made.
 type filteredExemplarReservoir[N int64 | float64] struct {
 	filter    exemplar.Filter
 	reservoir exemplar.Reservoir
+	lazyRes   lazyReservoir
 	// The exemplar.Reservoir is not required to be concurrent safe, but
 	// implementations can indicate that they are concurrent-safe by embedding
 	// reservoir.ConcurrentSafe in order to improve performance.
@@ -45,22 +50,38 @@ func NewFilteredExemplarReservoir[N int64 | float64](
 	r exemplar.Reservoir,
 ) FilteredExemplarReservoir[N] {
 	_, concurrentSafe := r.(reservoir.ConcurrentSafe)
+	lazyRes, _ := r.(lazyReservoir)
 	return &filteredExemplarReservoir[N]{
 		filter:         f,
 		reservoir:      r,
+		lazyRes:        lazyRes,
 		concurrentSafe: concurrentSafe,
 	}
 }
 
-func (f *filteredExemplarReservoir[N]) Offer(ctx context.Context, val N, attr []attribute.KeyValue) {
+func (f *filteredExemplarReservoir[N]) Offer(
+	ctx context.Context,
+	val N,
+	lazySet attribute.LazyFilteredSet,
+) {
 	if f.filter(ctx) {
 		// only record the current time if we are sampling this measurement.
 		ts := time.Now()
+		if f.lazyRes != nil {
+			if !f.concurrentSafe {
+				f.reservoirMux.Lock()
+				defer f.reservoirMux.Unlock()
+			}
+			f.lazyRes.OfferLazy(ctx, ts, exemplar.NewValue(val), lazySet)
+			return
+		}
+
+		dropped := lazySet.Dropped()
 		if !f.concurrentSafe {
 			f.reservoirMux.Lock()
 			defer f.reservoirMux.Unlock()
 		}
-		f.reservoir.Offer(ctx, ts, exemplar.NewValue(val), attr)
+		f.reservoir.Offer(ctx, ts, exemplar.NewValue(val), dropped)
 	}
 }