+++ sdk/metric/internal/aggregate/histogram.go
@@ -109,17 +109,16 @@ type deltaHistogram[N int64 | float64] struct {
 func (s *deltaHistogram[N]) measure(
 	ctx context.Context,
 	value N,
-	fltrAttr attribute.Set,
-	droppedAttr []attribute.KeyValue,
+	lazySet attribute.LazyFilteredSet,
 ) {
 	hotIdx := s.hcwg.start()
 	defer s.hcwg.done(hotIdx)
-	h := s.hotColdValMap[hotIdx].LoadOrStoreAttr(fltrAttr, func(attr attribute.Set) any {
-		r := s.newRes(attr)
+	h := s.hotColdValMap[hotIdx].LoadOrStoreAttr(lazySet, func(a attribute.Set) any {
+		r := s.newRes(a)
 		_, isDrop := r.(*dropRes[N])
-		hPt := &histogramPoint[N]{
+		return &histogramPoint[N]{
 			res:           r,
-			attrs:         attr,
+			attrs:         a,
 			dropExemplars: isDrop,
 			// N+1 buckets. For example:
 			//
@@ -130,7 +129,6 @@ func (s *deltaHistogram[N]) measure(
 			//   counts = (-∞, 0], (0, 5.0], (5.0, 10.0], (10.0, +∞)
 			histogramPointCounters: histogramPointCounters[N]{counts: make([]atomic.Uint64, len(s.bounds)+1)},
 		}
-		return hPt
 	}).(*histogramPoint[N])
 
 	// This search will return an index in the range [0, len(s.bounds)], where
@@ -147,7 +145,7 @@ func (s *deltaHistogram[N]) measure(
 		h.total.add(value)
 	}
 	if !h.dropExemplars {
-		h.res.Offer(ctx, value, droppedAttr)
+		h.res.Offer(ctx, value, lazySet)
 	}
 }
 
@@ -285,15 +283,14 @@ func newCumulativeHistogram[N int64 | float64](
 func (s *cumulativeHistogram[N]) measure(
 	ctx context.Context,
 	value N,
-	fltrAttr attribute.Set,
-	droppedAttr []attribute.KeyValue,
+	lazySet attribute.LazyFilteredSet,
 ) {
-	h := s.values.LoadOrStoreAttr(fltrAttr, func(attr attribute.Set) any {
-		r := s.newRes(attr)
+	h := s.values.LoadOrStoreAttr(lazySet, func(a attribute.Set) any {
+		r := s.newRes(a)
 		_, isDrop := r.(*dropRes[N])
-		hPt := &hotColdHistogramPoint[N]{
+		return &hotColdHistogramPoint[N]{
 			res:           r,
-			attrs:         attr,
+			attrs:         a,
 			startTime:     now(),
 			dropExemplars: isDrop,
 			// N+1 buckets. For example:
@@ -312,7 +309,6 @@ func (s *cumulativeHistogram[N]) measure(
 				},
 			},
 		}
-		return hPt
 	}).(*hotColdHistogramPoint[N])
 
 	// This search will return an index in the range [0, len(s.bounds)], where
@@ -333,7 +329,7 @@ func (s *cumulativeHistogram[N]) measure(
 		h.hotColdPoint[hotIdx].total.add(value)
 	}
 	if !h.dropExemplars {
-		h.res.Offer(ctx, value, droppedAttr)
+		h.res.Offer(ctx, value, lazySet)
 	}
 }